requirements.txt
# OSBot
git+https://github.com/owasp-sbot/OSBot-Utils.git@dev
git+https://github.com/owasp-sbot/OSBot-AWS.git@dev

# utils
python-dotenv

# aws
boto3

# for testing
pytest
coverage